chore: rename platform package to fleet_platform
Fixed naming conflict with Python standard library 'platform' module that was preventing uvicorn from starting. All 16 tests passing, server smoke tests verified.
