+++ fleet_platform/db/session.py
@@ -2,7 +2,7 @@ from collections.abc import AsyncGenerator
 
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 
-from platform.core.config import settings
+from fleet_platform.core.config import settings
 
 engine = create_async_engine(
     settings.database_url,
