+++ fleet_platform/models/base.py
