+++ fleet_platform/models/drift.py
@@ -5,7 +5,7 @@ from sqlalchemy import BigInteger, DateTime, ForeignKey, Index, Integer, SmallIn
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
-from platform.models.base import Base, TimestampMixin
+from fleet_platform.models.base import Base, TimestampMixin
 
 
 class DesiredStateBaseline(Base, TimestampMixin):
