+++ fleet_platform/models/execution.py
@@ -5,7 +5,7 @@ from sqlalchemy import DateTime, ForeignKey, Index, Integer, String, Text
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
-from platform.models.base import Base
+from fleet_platform.models.base import Base
 
 
 class ExecutionJob(Base):
