--- platform/models/__init__.py
@@ -1,21 +1,0 @@
-from platform.models.base import Base, TimestampMixin
-from platform.models.user import User
-from platform.models.node import Node, Tag
-from platform.models.group import Group, GroupMember
-from platform.models.facts import NodeFact
-from platform.models.drift import DesiredStateBaseline, DriftRecord
-from platform.models.sbom import SBOMScan, SBOMComponent
-from platform.models.execution import ExecutionJob, ExecutionResult
-from platform.models.audit import AuditEvent
-
-__all__ = [
-    "Base", "TimestampMixin",
-    "User",
-    "Node", "Tag",
-    "Group", "GroupMember",
-    "NodeFact",
-    "DesiredStateBaseline", "DriftRecord",
-    "SBOMScan", "SBOMComponent",
-    "ExecutionJob", "ExecutionResult",
-    "AuditEvent",
-]
