+++ pyproject.toml
@@ -38,4 +38,4 @@ target-version = "py313"
 select = ["E", "F", "I"]
 
 [tool.ruff.lint.isort]
-known-first-party = ["platform"]
+known-first-party = ["fleet_platform"]
